stress-ioprio: cast ioprio_get who arg to int
Clean up pedantic clang warning, cast uid to int

+++ stress-ioprio.c
@@ -121,7 +121,7 @@ static int stress_ioprio(const stress_args_t *args)
 		(void)shim_ioprio_get(IOPRIO_WHO_PGRP, ~0);
 		(void)shim_ioprio_get(IOPRIO_WHO_USER, ~0);
 
-		if (shim_ioprio_get(IOPRIO_WHO_USER, uid) < 0) {
+		if (shim_ioprio_get(IOPRIO_WHO_USER, (int)uid) < 0) {
 			if (errno != EINVAL) {
 				pr_fail("%s: ioprio_get(OPRIO_WHO_USR, %d), "
 					"errno = %d (%s)\n",
